state: rename applyTx.js
