+++ packages/portalnetwork/src/networks/state/applyTx.ts
