+++ packages/portalnetwork/src/networks/state/index.ts
@@ -1,6 +1,6 @@
+export * from './applyTx.js'
 export * from './nibbleEncoding.js'
 export * from './routingTable.js'
-export * from './runBlock.js'
 export * from './state.js'
 export * from './stateManager.js'
 export * from './types.js'
