+++ packages/portalnetwork/src/networks/state/state.ts
@@ -20,8 +20,8 @@ import { decodeHistoryNetworkContentKey } from '../history/util.js'
 import { BaseNetwork } from '../network.js'
 import { NetworkId } from '../types.js'
 
+import { applyTransactions } from './applyTx.js'
 import { addressToNibbles, packNibbles, unpackNibbles } from './nibbleEncoding.js'
-import { applyTransactions } from './runBlock.js'
 import { StateDB } from './statedb.js'
 import { AccountTrieNodeOffer, AccountTrieNodeRetrieval, StateNetworkContentType } from './types.js'
 import { AccountTrieNodeContentKey, StateNetworkContentId, nextOffer } from './util.js'
